requirements.txt
asgiref==3.9.2
Django==5.2.6
djangorestframework==3.16.1
pillow==11.3.0
sqlparse==0.5.3
tzdata==2025.2
python-jose==3.5.0
cryptography==46.0.3
